+++ src/main/java/org/gitlab4j/api/PackagesApi.java
@@ -52,7 +52,7 @@ public class PackagesApi extends AbstractApi {
      *
      * <pre><code>GitLab Endpoint: GET /projects/:id/packages</code></pre>
      *
-     * @param projectIdOrPath the project in the form of an Integer(ID), String(path), or Project instance
+     * @param projectIdOrPath the project in the form of an Long(ID), String(path), or Project instance
      * @return a list of pages in the project's packages
      * @throws GitLabApiException if any exception occurs
      */
@@ -66,7 +66,7 @@ public class PackagesApi extends AbstractApi {
      *
      * <pre><code>GitLab Endpoint: GET /projects/:id/packages</code></pre>
      *
-     * @param projectIdOrPath the project in the form of an Integer(ID), String(path), or Project instance
+     * @param projectIdOrPath the project in the form of an Long(ID), String(path), or Project instance
      * @param page the page to get
      * @param perPage the number of Package instances per page
      * @return a list of project packages for the specified range
@@ -84,7 +84,7 @@ public class PackagesApi extends AbstractApi {
      *
      * <pre><code>GitLab Endpoint: GET /projects/:id/packages</code></pre>
      *
-     * @param projectIdOrPath the project in the form of an Integer(ID), String(path), or Project instance
+     * @param projectIdOrPath the project in the form of an Long(ID), String(path), or Project instance
      * @param itemsPerPage the number of Package instances per page
      * @return a Pager of project packages for the specified range
      * @throws GitLabApiException if any exception occurs
@@ -99,7 +99,7 @@ public class PackagesApi extends AbstractApi {
      *
      * <pre><code>GitLab Endpoint: GET /projects/:id/packages</code></pre>
      *
-     * @param projectIdOrPath the project in the form of an Integer(ID), String(path), or Project instance
+     * @param projectIdOrPath the project in the form of an Long(ID), String(path), or Project instance
      * @param filter the PackageFilter instance holding the filter values for the query
      * @param itemsPerPage the number of Package instances per page
      * @return a Pager of project packages for the specified range
@@ -117,7 +117,7 @@ public class PackagesApi extends AbstractApi {
      *
      * <pre><code>GitLab Endpoint: GET /projects/:id/packages</code></pre>
      *
-     * @param projectIdOrPath the project in the form of an Integer(ID), String(path), or Project instance
+     * @param projectIdOrPath the project in the form of an Long(ID), String(path), or Project instance
      * @return a Stream of pages in the project's packages
      * @throws GitLabApiException if any exception occurs
      */
@@ -131,7 +131,7 @@ public class PackagesApi extends AbstractApi {
      *
      * <pre><code>GitLab Endpoint: GET /projects/:id/packages</code></pre>
      *
-     * @param projectIdOrPath the project in the form of an Integer(ID), String(path), or Project instance
+     * @param projectIdOrPath the project in the form of an Long(ID), String(path), or Project instance
      * @param filter the PackageFilter instance holding the filter values for the query
      * @return a Stream of pages in the project's packages
      * @throws GitLabApiException if any exception occurs
@@ -145,12 +145,12 @@ public class PackagesApi extends AbstractApi {
      *
      * <pre><code>GitLab Endpoint: GET /projects/:id/packages/:package_id</code></pre>
      *
-     * @param projectIdOrPath the project in the form of an Integer(ID), String(path), or Project instance
+     * @param projectIdOrPath the project in the form of an Long(ID), String(path), or Project instance
      * @param packageId the ID of the package to get
      * @return a Package instance for the specified package ID
      * @throws GitLabApiException if any exception occurs
      */
-    public Package getPackage(Object projectIdOrPath, Integer packageId) throws GitLabApiException {
+    public Package getPackage(Object projectIdOrPath, Long packageId) throws GitLabApiException {
         Response response = get(Response.Status.OK, null,
                 "projects", getProjectIdOrPath(projectIdOrPath), "packages", packageId);
         return (response.readEntity(Package.class));
@@ -161,12 +161,12 @@ public class PackagesApi extends AbstractApi {
      *
      * <pre><code>GitLab Endpoint: GET /projects/:id/packages/:package_id/package_files</code></pre>
      *
-     * @param projectIdOrPath the project in the form of an Integer(ID), String(path), or Project instance
+     * @param projectIdOrPath the project in the form of an Long(ID), String(path), or Project instance
      * @param packageId the ID of the package to get the package files for
      * @return a list of PackageFile instances for the specified package ID
      * @throws GitLabApiException if any exception occurs
      */
-    public List<PackageFile> getPackageFiles(Object projectIdOrPath, Integer packageId) throws GitLabApiException {
+    public List<PackageFile> getPackageFiles(Object projectIdOrPath, Long packageId) throws GitLabApiException {
         return (getPackageFiles(projectIdOrPath, packageId, getDefaultPerPage()).all());
     }
 
@@ -175,14 +175,14 @@ public class PackagesApi extends AbstractApi {
      *
      * <pre><code>GitLab Endpoint: GET /projects/:id/packages/:package_id/package_files</code></pre>
      *
-     * @param projectIdOrPath the project in the form of an Integer(ID), String(path), or Project instance
+     * @param projectIdOrPath the project in the form of an Long(ID), String(path), or Project instance
      * @param packageId the ID of the package to get the package files for
      * @param page the page to get
      * @param perPage the number of PackageFile instances per page
      * @return a list of PackageFile instances for the specified package ID
      * @throws GitLabApiException if any exception occurs
      */
-    public List<PackageFile> getPackageFiles(Object projectIdOrPath, Integer packageId, int page, int perPage) throws GitLabApiException {
+    public List<PackageFile> getPackageFiles(Object projectIdOrPath, Long packageId, int page, int perPage) throws GitLabApiException {
         Response response = get(Response.Status.OK, getPageQueryParams(page, perPage),
                 "projects", getProjectIdOrPath(projectIdOrPath), "packages", packageId, "package_files");
         return response.readEntity(new GenericType<List<PackageFile>>() {});
@@ -193,13 +193,13 @@ public class PackagesApi extends AbstractApi {
      *
      * <pre><code>GitLab Endpoint: GET /projects/:id/packages/:package_id/package_files</code></pre>
      *
-     * @param projectIdOrPath the project in the form of an Integer(ID), String(path), or Project instance
+     * @param projectIdOrPath the project in the form of an Long(ID), String(path), or Project instance
      * @param packageId the ID of the package to get the package files for
      * @param itemsPerPage the number of PackageFile instances per page
      * @return a Pager of PackageFile instances for the specified package ID
      * @throws GitLabApiException if any exception occurs
      */
-    public Pager<PackageFile> getPackageFiles(Object projectIdOrPath,  Integer packageId, int itemsPerPage) throws GitLabApiException {
+    public Pager<PackageFile> getPackageFiles(Object projectIdOrPath,  Long packageId, int itemsPerPage) throws GitLabApiException {
         return (new Pager<PackageFile>(this, PackageFile.class, itemsPerPage, null,
                 "projects", getProjectIdOrPath(projectIdOrPath), "packages", packageId, "package_files"));
     }
@@ -209,12 +209,12 @@ public class PackagesApi extends AbstractApi {
      *
      * <pre><code>GitLab Endpoint: GET /projects/:id/packages/:package_id/package_files</code></pre>
      *
-     * @param projectIdOrPath the project in the form of an Integer(ID), String(path), or Project instance
+     * @param projectIdOrPath the project in the form of an Long(ID), String(path), or Project instance
      * @param packageId the ID of the package to get the package files for
      * @return a Stream of PackageFile instances for the specified package ID
      * @throws GitLabApiException if any exception occurs
      */
-    public Stream<PackageFile> getPackagesStream(Object projectIdOrPath, Integer packageId) throws GitLabApiException {
+    public Stream<PackageFile> getPackagesStream(Object projectIdOrPath, Long packageId) throws GitLabApiException {
         return (getPackageFiles(projectIdOrPath, packageId, getDefaultPerPage()).stream());
     }
 
@@ -223,11 +223,11 @@ public class PackagesApi extends AbstractApi {
      *
      * <pre><code>GitLab Endpoint: DELETE /projects/:id/packages/:package_id</code></pre>
      *
-     * @param projectIdOrPath the project in the form of an Integer(ID), String(path), or Project instance
+     * @param projectIdOrPath the project in the form of an Long(ID), String(path), or Project instance
      * @param packageId the ID of the package to delete
      * @throws GitLabApiException if any exception occurs
      */
-    public void deletePackage(Object projectIdOrPath, Integer packageId) throws GitLabApiException {
+    public void deletePackage(Object projectIdOrPath, Long packageId) throws GitLabApiException {
 
         if (packageId == null) {
             throw new RuntimeException("packageId cannot be null");
