+++ src/main/java/org/gitlab4j/api/RunnersApi.java
@@ -265,7 +265,7 @@ public class RunnersApi extends AbstractApi {
      * @return RunnerDetail instance.
      * @throws GitLabApiException if any exception occurs
      */
-    public RunnerDetail getRunnerDetail(Integer runnerId) throws GitLabApiException {
+    public RunnerDetail getRunnerDetail(Long runnerId) throws GitLabApiException {
 
         if (runnerId == null) {
             throw new RuntimeException("runnerId cannot be null");
@@ -290,7 +290,7 @@ public class RunnersApi extends AbstractApi {
      * @return RunnerDetail instance.
      * @throws GitLabApiException if any exception occurs
      */
-    public RunnerDetail updateRunner(Integer runnerId, String description, Boolean active, List<String> tagList,
+    public RunnerDetail updateRunner(Long runnerId, String description, Boolean active, List<String> tagList,
                                      Boolean runUntagged, Boolean locked, RunnerDetail.RunnerAccessLevel accessLevel) throws GitLabApiException {
         if (runnerId == null) {
             throw new RuntimeException("runnerId cannot be null");
@@ -315,7 +315,7 @@ public class RunnersApi extends AbstractApi {
      * @param runnerId The ID of a runner
      * @throws GitLabApiException if any exception occurs
      */
-    public void removeRunner(Integer runnerId) throws GitLabApiException {
+    public void removeRunner(Long runnerId) throws GitLabApiException {
 
         if (runnerId == null) {
             throw new RuntimeException("runnerId cannot be null");
@@ -333,7 +333,7 @@ public class RunnersApi extends AbstractApi {
      * @return List jobs that are being processed or were processed by specified Runner
      * @throws GitLabApiException if any exception occurs
      */
-    public List<Job> getJobs(Integer runnerId) throws GitLabApiException {
+    public List<Job> getJobs(Long runnerId) throws GitLabApiException {
         return (getJobs(runnerId, null, getDefaultPerPage()).all());
     }
 
@@ -346,7 +346,7 @@ public class RunnersApi extends AbstractApi {
      * @return a Stream of jobs that are being processed or were processed by specified Runner
      * @throws GitLabApiException if any exception occurs
      */
-    public Stream<Job> getJobsStream(Integer runnerId) throws GitLabApiException {
+    public Stream<Job> getJobsStream(Long runnerId) throws GitLabApiException {
         return (getJobs(runnerId, null, getDefaultPerPage()).stream());
     }
 
@@ -360,7 +360,7 @@ public class RunnersApi extends AbstractApi {
      * @return List jobs that are being processed or were processed by specified Runner
      * @throws GitLabApiException if any exception occurs
      */
-    public List<Job> getJobs(Integer runnerId, JobStatus status) throws GitLabApiException {
+    public List<Job> getJobs(Long runnerId, JobStatus status) throws GitLabApiException {
         return (getJobs(runnerId, status, getDefaultPerPage()).all());
     }
 
@@ -374,7 +374,7 @@ public class RunnersApi extends AbstractApi {
      * @return a Stream of jobs that are being processed or were processed by specified Runner
      * @throws GitLabApiException if any exception occurs
      */
-    public Stream<Job> getJobsStream(Integer runnerId, JobStatus status) throws GitLabApiException {
+    public Stream<Job> getJobsStream(Long runnerId, JobStatus status) throws GitLabApiException {
         return (getJobs(runnerId, status, getDefaultPerPage()).stream());
     }
 
@@ -388,7 +388,7 @@ public class RunnersApi extends AbstractApi {
      * @return a Pager containing the Jobs for the Runner
      * @throws GitLabApiException if any exception occurs
      */
-    public Pager<Job> getJobs(Integer runnerId, int itemsPerPage) throws GitLabApiException {
+    public Pager<Job> getJobs(Long runnerId, int itemsPerPage) throws GitLabApiException {
         return (getJobs(runnerId, null, itemsPerPage));
     }
 
@@ -403,7 +403,7 @@ public class RunnersApi extends AbstractApi {
      * @return a Pager containing the Jobs for the Runner
      * @throws GitLabApiException if any exception occurs
      */
-    public Pager<Job> getJobs(Integer runnerId, JobStatus status, int itemsPerPage) throws GitLabApiException {
+    public Pager<Job> getJobs(Long runnerId, JobStatus status, int itemsPerPage) throws GitLabApiException {
 
         if (runnerId == null) {
             throw new RuntimeException("runnerId cannot be null");
@@ -419,7 +419,7 @@ public class RunnersApi extends AbstractApi {
      *
      * <pre><code>GitLab Endpoint: GET /projects/:id/runners</code></pre>
      *
-     * @param projectIdOrPath the project in the form of an Integer(ID), String(path), or Project instance
+     * @param projectIdOrPath the project in the form of an Long(ID), String(path), or Project instance
      * @return List of all Runner available in the project
      * @throws GitLabApiException if any exception occurs
      */
@@ -433,7 +433,7 @@ public class RunnersApi extends AbstractApi {
      *
      * <pre><code>GitLab Endpoint: GET /projects/:id/runners</code></pre>
      *
-     * @param projectIdOrPath the project in the form of an Integer(ID), String(path), or Project instance
+     * @param projectIdOrPath the project in the form of an Long(ID), String(path), or Project instance
      * @return a Stream of all Runner available in the project
      * @throws GitLabApiException if any exception occurs
      */
@@ -447,7 +447,7 @@ public class RunnersApi extends AbstractApi {
      *
      * <pre><code>GitLab Endpoint: GET /projects/:id/runners</code></pre>
      *
-     * @param projectIdOrPath the project in the form of an Integer(ID), String(path), or Project instance
+     * @param projectIdOrPath the project in the form of an Long(ID), String(path), or Project instance
      * @param itemsPerPage the number of Project instances that will be fetched per page
      * @return Pager of all Runner available in the project
      * @throws GitLabApiException if any exception occurs
@@ -462,12 +462,12 @@ public class RunnersApi extends AbstractApi {
      *
      * <pre><code>GitLab Endpoint: POST /projects/:id/runners</code></pre>
      *
-     * @param projectIdOrPath the project in the form of an Integer(ID), String(path), or Project instance
+     * @param projectIdOrPath the project in the form of an Long(ID), String(path), or Project instance
      * @param runnerId  The ID of a runner
      * @return Runner instance of the Runner enabled
      * @throws GitLabApiException if any exception occurs
      */
-    public Runner enableRunner(Object projectIdOrPath, Integer runnerId) throws GitLabApiException {
+    public Runner enableRunner(Object projectIdOrPath, Long runnerId) throws GitLabApiException {
         GitLabApiForm formData = new GitLabApiForm().withParam("runner_id", runnerId, true);
         Response response = post(Response.Status.CREATED, formData.asMap(),
                 "projects", getProjectIdOrPath(projectIdOrPath), "runners");
@@ -480,12 +480,12 @@ public class RunnersApi extends AbstractApi {
      *
      * <pre><code>GitLab Endpoint: DELETE /projects/:id/runners/:runner_id</code></pre>
      *
-     * @param projectIdOrPath the project in the form of an Integer(ID), String(path), or Project instance
+     * @param projectIdOrPath the project in the form of an Long(ID), String(path), or Project instance
      * @param runnerId  The ID of a runner
      * @return Runner instance of the Runner disabled
      * @throws GitLabApiException if any exception occurs
      */
-    public Runner disableRunner(Object projectIdOrPath, Integer runnerId) throws GitLabApiException {
+    public Runner disableRunner(Object projectIdOrPath, Long runnerId) throws GitLabApiException {
         Response response = delete(Response.Status.OK, null,
                 "projects", getProjectIdOrPath(projectIdOrPath), "runners", runnerId);
         return (response.readEntity(Runner.class));
