+++ src/main/java/org/gitlab4j/api/models/Membership.java
@@ -4,16 +4,16 @@ import org.gitlab4j.api.utils.JacksonJson;
 
 public class Membership {
 
-    private Integer sourceId;
+    private Long sourceId;
     private String sourceName;
     private MembershipSourceType sourceType;
     private AccessLevel accessLevel;
 
-    public Integer getSourceId() {
+    public Long getSourceId() {
         return sourceId;
     }
 
-    public void setSourceId(Integer sourceId) {
+    public void setSourceId(Long sourceId) {
         this.sourceId = sourceId;
     }
